backend/requirements.txt
flask
flask-cors
flask-limiter
flask-login
flask_mail_sendgrid
flask-sqlalchemy
werkzeug
python-dotenv
gunicorn
markdown
nltk
faiss-cpu
sentence-transformers
langchain
langchain-community
openai
pymupdf
email-validator
pillow
python-dateutil
jinja2
